+++ xcube/plugin.py
@@ -191,6 +191,7 @@ def _register_cli_commands(ext_registry: extension.ExtensionRegistry):
         'tile',
         'vars2dim',
         'verify',
+        'versions',
 
         # Experimental + Hidden
         'io',