+++ CHANGES.md
@@ -83,9 +83,10 @@
 * Added a new utility module `xcube.util.temp` that allows for creating 
   temporary files and directories that will be deleted when the current 
   process ends.
-* Added `xcube.util.versions.XCUBE_VERSIONS` and function 
-  `xcube.util.versions.get_versions` to get mappings of package names to 
-  package versions. (#522)
+* Added function `xcube.util.versions.get_xcube_versions()`  
+  that outputs the versions of packages relevant for xcube.
+  Also added a new CLI `xcube versions` that outputs the result of the  
+  new function in JSON or YAML. (#522)
 
 ### Other